Outer interpreter debugged
type "r", "g", "b" to turn on the red, green and blue LEDs respectively. The outer interpreter looks up the character in the dictionary and returns the address to be executed by the "execute" opcode.

Pay attention to RESET the computer before trying this program since it is critical for correct operation. This was the cause of the "bug" encountered where the program was unresponsive.