Words as strings
Words can now be strings such as "red", "green" and "blue". They are represented in the dictionary as a count and three characters. This allows an extremely fast search but may not allow enough characters to tell words apart. This is a starting point that may be expanded on. The dictionary representation and search should be abstracted.